requirements.txt
# Google API dependencies
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
gspread>=5.0.0
oauth2client>=4.1.0

# Production dependencies
requests>=2.25.0
python-dotenv>=0.19.0

# Optional: Web framework for API endpoints (if needed)
# flask>=2.0.0
# gunicorn>=20.1.0 